requirements.txt
fastapi
uvicorn
python-multipart
gspread
oauth2client
python-dotenv
google-genai
Pillow
requests
